Rename UserController and UserModel
Rename controllers/users into controllers/UserController
and models/User into models/UserModel so that it will remove names
duplicates when using search tools.
