+++ sources/api/routes/users/index.ts
@@ -5,7 +5,7 @@
 import { Router } from "express";
 import * as AsyncHandler from 'express-async-handler';
 
-import * as UserController from '../../controllers/users';
+import * as UserController from '../../controllers/UserController';
 
 const router = Router();
